refactor(dir): Use subdir to avoid name conflicts
